fix(activerecord): wire cast type into SqlTypeMetadata; fix FK test parent_id type
- Pass castType.type() from TypeMap OID lookup into SqlTypeMetadata when
  constructing PG columns, mirroring Rails fetch_type_metadata which sets
  type: cast_type.type. Fixes Column#isEnum returning false for user-defined
  enums (OID::Enum#type returns "enum" but SqlTypeMetadata defaulted type to
  the raw sqlType string like "mood").
- Change parent_id from INT to BIGINT in referential-integrity test to match
  id BIGSERIAL (bigint), which PG requires for FK column type compatibility.

+++ packages/activerecord/src/adapters/postgresql/referential-integrity.test.ts
@@ -32,7 +32,7 @@ describeIfPg("PostgreSQLAdapter", () => {
       await adapter.execute(`
         CREATE TABLE referential_integrity_test_schema.nodes (
           id        BIGSERIAL,
-          parent_id INT NOT NULL,
+          parent_id BIGINT NOT NULL,
           PRIMARY KEY (id),
           CONSTRAINT fk_parent_node FOREIGN KEY (parent_id)
             REFERENCES referential_integrity_test_schema.nodes (id)
